// src/app/api/model/sectionDTO.ts
/**
 * FastAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface SectionDTO { 
    number: number;
    name: string;
    state?: string | null;
    district?: string | null;
}